Both LibreOffice and Inkscape are still GTK3 // Update README.md
Both LibreOffice and Inkscape are still GTK3:

LibreOffice is still GTK3. GTK4 is in development with no clear timeframe. The GTK4 port is not yet finished.

Inkscape just released v1.4 on basis of GTK3. The next release v1.5 will be based on GTK4, but there is no alpha or beta yet.