Slight improvement to the description of "Bottles"
I added "to run Windows software on Linux" to the description of "Bottles".

(I don't think that Bottles is in the correct section (Development -> Containers). Bottles is an end user application to emulate Windows software via Wine. It is neither development nor container. However, there is no section, where it would fit better.)